web+api: fold Packs into Capabilities (subsection + env-on-enable)
Make the generic capability-enable path the single way to enable a pack,
and fold the standalone Packs page into a first-class subsection of
Capabilities.

API (env-on-enable):
- EnableCapabilityRequest gains optional environmentId (contracts + SDK type).
- POST /capabilities/:id/enable threads the full grant into enableCapability.
- enableCapability's pack branch now accepts an initial environment
  attachment: a request-supplied environmentId is validated as a fresh
  attachment (validateEnvironmentAttachment + required-variables check) and
  persisted on the pack installation, mirroring POST /packs/:id/enable. With
  no request id the previously-stored attachment is re-validated; an
  environment.required pack with neither still 422s with a clear message.
  assertPackSandboxImageCompatible and the one-image-pack-per-workspace 409
  invariant are untouched.

Web (IA):
- Delete the standalone routes/packs.tsx; render Packs as a subsection at the
  top of the Capabilities catalog (register-manifest / unregister /
  enable-with-environment / disable preserved). Generic CapabilityRow rows
  exclude kind===pack so packs never double-render.
- Pack enable now rides the unified client.enableCapability(pack:{id}, { environmentId }).
- /workspaces/:id/packs redirects to /capabilities?section=packs (Packs
  subsection focused); the separate Packs nav tab is removed.

Tests: integration covers env-on-enable through the unified path (201 +
stored attachment, bogus id 422, missing id 422); the sandbox-image 409
invariant test is intact.
